move cmds to specific directories
