+++ commands/sgprot/sgplist.js
@@ -1,5 +1,5 @@
-const sgprot = require('../sgprot.json')
-const tools = require('../tools.js')
+const sgprot = require('../../sgprot.json')
+const tools = require('../../tools.js')
 
 module.exports.run = async (ts, ev, client) => {
   if (tools.isEmpty(sgprot)) return ts.sendTextMessage(client.getID(), 1, 'There are no protected groups.')
